requirements/default.txt
boto3==1.7.4 \
    --hash=sha256:13ad5f64a247d655a27dca83274588e9d14cba61b38d3d4fd2b011b7197d88dd \
    --hash=sha256:a56b21efbc994580fc9cef454f0f949745c152326f939aed6609d1c47b2a0f8f
Django==2.0.4 \
    --hash=sha256:2d8b9eed8815f172a8e898678ae4289a5e9176bc08295676eff4228dd638ea61 \
    --hash=sha256:d81a1652963c81488e709729a80b510394050e312f386037f26b54912a3a10d0
python-dotenv==0.8.2 \
    --hash=sha256:4965ed170bf51c347a89820e8050655e9c25db3837db6602e906b6d850fad85c \
    --hash=sha256:509736185257111613009974e666568a1b031b028b61b500ef1ab4ee780089d5
click==6.7 \
    --hash=sha256:29f99fc6125fbc931b758dc053b3114e55c77a6e4c6c3a2674a2dc986016381d \
    --hash=sha256:f15516df478d5a56180fbf80e68f206010e6d160fc39fa508b65e035fd75130b
raven==6.7.0 \
    --hash=sha256:e4edf648829a64234800a10ed94ca08e0b38592f7449fa5e70931db62f5cd851 \
    --hash=sha256:f908e9b39f02580e7f822030d119ed3b2e8d32300a2fec6373e5827d588bbae7
